requirements.txt
click
python-dotenv
bencodepy